Add hover and new grid generation functionality
Added a mouseover eventListener to change the
color of elements that are hovered over in the grid.
Added a button with input for a new grid size and
refactored the grid generation logic into a function.
Additionally used replaceChildren to remove the previous
appended child elements and empty the main container.